+++ src/WS1S/mona-1.4/DFA/product.c
@@ -58,12 +58,12 @@ unsigned prod_term_fn(unsigned  p, unsigned q)
 {    
   int res;
 
-  if ( (res = (int) lookup_in_hash_tab(htbl, p, q)) )
+  if ( (res = (int) (long) lookup_in_hash_tab(htbl, p, q)) )
     /* res = 0 or id+1 */
     return (--res);
   else {
     insert_in_hash_tab(htbl,  p, 
-		       q, (void *) (res = ++last_state));
+		       q, (void *) (long) (res = ++last_state));
     qt->next = new_list(p, q, (list) 0);
     qt = qt->next;
 
@@ -75,7 +75,7 @@ unsigned prod_term_fn(unsigned  p, unsigned q)
 /*insert a loop for the product state (p, q) */
 inline void make_loop (bdd_manager *bddm, unsigned  p, unsigned q) {
   int res;
-  res = (int) lookup_in_hash_tab(htbl, p, q);
+  res = (int) (long) lookup_in_hash_tab(htbl, p, q);
   invariant(res);
   /* res = 0 or id+1 */
   (--res);
